requirements.txt
torch
torchvision
numpy
pandas
pillow
scikit-learn
tqdm
matplotlib
